Removes unecessary file
